# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Self Deepsearch Platform API
  version: 0.1.0
  description: Release A contract baseline. Automatic collection is not enabled.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/auth/logout:
    post:
      operationId: logout
      description: Idempotently revokes the current session. Only confirmed revocation or an absent session clears the session and recent-auth cookies. Dependency failures leave cookies unchanged so the caller can retry; they never imply successful logout.
      responses:
        "204":
          description: Current session revoked or already absent
        "503":
          description: Session revocation could not be confirmed; cookies remain unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/me/favorites:
    get:
      operationId: listFavorites
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Visible favorite works
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
  /api/v1/me/favorites/{work_id}:
    parameters:
      - in: path
        name: work_id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: addFavorite
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Favorite is visible }
    delete:
      operationId: removeFavorite
      description: Soft deletes the favorite.
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Favorite is no longer visible }
  /api/v1/me/follows:
    get:
      operationId: listFollows
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Visible followed performers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
  /api/v1/me/followed-works:
    get:
      operationId: listFollowedWorks
      description: Published works from followed performers, excluding the user's active hidden preferences.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: New works from followed performers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemListResponse" }
        "401": { description: Login required }
  /api/v1/me/follows/{performer_id}:
    parameters:
      - in: path
        name: performer_id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: addFollow
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Follow is visible }
    delete:
      operationId: removeFollow
      description: Soft deletes the follow.
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Follow is no longer visible }
  /api/v1/me/hidden-works:
    get:
      operationId: listHiddenWorks
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Published works hidden from this account's discovery surfaces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
...
  /api/v1/me/history:
    get:
      operationId: listHistory
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Visible precise history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
    delete:
      operationId: clearVisibleHistory
      description: Sets is_deleted; retained records are hidden and may be archived later.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Number of records hidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearHistoryResponse"
  /api/v1/feedback:
    post:
      operationId: createFeedback
      description: Stores text and an optional URL without fetching the URL.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "201":
          description: Feedback queued for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackItem"
...
  /admin/v1/feedback:
    get:
      operationId: adminListFeedback
      description: Lists feedback for manual review. Requires editor, admin or owner.
      security: [{ sessionCookie: [] }]
      parameters:
        - in: query
          name: status
          schema: { enum: [all, pending, reviewing, accepted, rejected, closed], default: all }
      responses:
        "200":
          description: Feedback review queue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFeedbackListResponse" }
        "403": { description: Role is not allowed }
  /admin/v1/feedback/{feedback_id}/review:
    parameters:
      - in: path
        name: feedback_id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: adminReviewFeedback
      description: Editors may move pending feedback to reviewing; admin or owner may finish review.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackReviewRequest" }
      responses:
        "200":
          description: Updated feedback
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFeedbackItem" }
        "409": { description: Feedback state changed }
...
components:
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sd_session
    recentAuthCookie:
      type: apiKey
      in: cookie
      name: sd_reauth
...
  schemas:
...
    HomeItem:
      type: object
      required: [id, type, title, subtitle, href, image_url, image]
      properties:
        id:
          type: string
          format: uuid
        type:
          enum: [work, performer, studio]
        code:
          type: string
        title:
          type: string
        subtitle:
          type: string
        href:
          type: string
        image_url:
          type: [string, "null"]
          description: Backward-compatible preferred rendition URL. New clients should use image.renditions.
        image:
          oneOf:
            - $ref: "#/components/schemas/ResponsiveImage"
            - type: "null"
...
    ImageRendition:
      type: object
      required: [url, rendition, width, height, mime_type]
      properties:
        url:
          type: string
          format: uri
        rendition:
          enum: [w320, w640, w960]
        width:
          type: integer
        height:
          type: integer
        mime_type:
          enum: [image/jpeg, image/png, image/webp]
    ResponsiveImage:
      allOf:
        - $ref: "#/components/schemas/ImageRendition"
        - type: object
          required: [renditions]
          properties:
            renditions:
              type: array
              minItems: 1
              maxItems: 3
              uniqueItems: true
              description: Available public derivatives for the same logical media asset; URLs are returned by the API and must not be inferred.
              items:
                $ref: "#/components/schemas/ImageRendition"
...
    ItemListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/HomeItem" }
        next_cursor: { type: [string, "null"], maxLength: 256 }
    HistoryItem:
      type: object
      required: [item, viewed_at]
      properties:
        item: { $ref: "#/components/schemas/HomeItem" }
        viewed_at: { type: string, format: date-time }
    HistoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/HistoryItem" }
    ClearHistoryResponse:
      type: object
      required: [hidden_count]
      properties:
        hidden_count: { type: integer, minimum: 0 }
    FeedbackRequest:
      type: object
      additionalProperties: false
      required: [feedback_type, message]
      properties:
        content_type: { type: [string, "null"], enum: [work, performer, studio, null] }
        content_id: { type: [string, "null"], format: uuid }
        feedback_type: { enum: [correction, source_suggestion, rights, other] }
        message: { type: string, minLength: 1, maxLength: 4000 }
        evidence_url: { type: [string, "null"], format: uri, maxLength: 2048 }
    FeedbackItem:
      allOf:
        - $ref: "#/components/schemas/FeedbackRequest"
        - type: object
          required: [id, content_type, content_id, evidence_url, review_status, created_at]
          properties:
            id: { type: string, format: uuid }
            review_status: { enum: [pending, reviewing, accepted, rejected, closed] }
            created_at: { type: string, format: date-time }
...
    AdminFeedbackItem:
      allOf:
        - $ref: "#/components/schemas/FeedbackItem"
        - type: object
          required: [submitter, reviewer, reviewed_at]
          properties:
            submitter: { type: string, format: email }
            reviewer: { type: [string, "null"], format: email }
            reviewed_at: { type: [string, "null"], format: date-time }
    AdminFeedbackListResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AdminFeedbackItem" } }
    FeedbackReviewRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { enum: [reviewing, accepted, rejected, closed] }
        reason: { type: string, minLength: 2, maxLength: 1000 }
...